stream-data-machine/requirements.txt
apache-airflow==2.9.1
alembic==1.13.1
argcomplete==3.3.0
pandas==2.2.2
